:rotating_light: fix the state initialization issue #49 (#52)
Some states should've survived the activity or process recreation. For example, it should've survived when the screen is rotated in the Android application.
But these states were declared as 'remember'. So I fixed it to 'rememberSaveable'. By doing this, these states will survive on those situations.

After fixing this issue, another issue arised. It's related with OrderedTask.
In this function, it modifies each step's 'onComplete' function. This process must be called only once.
But when the screen is rotated, this code called again and again.
To fix this issue, I modified to call that process only once at the first time.

Signed-off-by: culater.kwak <culater.kwak@samsung.com>